requirements.txt
streamlit
torch
transformers
Pillow
ultralytics